Remove namespace from server size URLs, add permissions to ServerSizeViewset, unit tests.
